boardmember component added
