+++ src/sections/about/Teammate.jsx
@@ -1,40 +1,65 @@
-import React, { Component } from "react";
+import React, { useEffect } from 'react';
 
 // Stylesheets
 import "./stylesheets/OurTeamStyles.scss";
+import Axios from 'axios';
 
-export class Teammate extends Component {
-  render() {
-    return (
-      <div>
-        <img
-          src="https://theshortcut.org/wp-content/uploads/2019/09/Our-Team-aug19_Anne.png"
-          alt="teammate"
-        />
+export default function Teammate() {
+  const [teammates, setTeammates] = React.useState([]);
+  const URL =
+    "https://theshortcut.org/wp-json/wp/v2/team/?per_page=100";
 
-        <div className="text">
-          <h4>Anne Badan</h4>
-          <p>Co-founder & CEO</p>
-        </div>
 
-        <div className="social-links">
-          <div>
-            <img
-              src="https://theshortcut.org/wp-content/uploads/2019/10/email-icon-01.png"
-              alt="mail-icon"
-            ></img>
-          </div>
-          <div>
-            {" "}
-            <img
-              src="https://theshortcut.org/wp-content/uploads/2019/10/linkedin-2.png"
-              alt="linkedin-icon"
-            />
-          </div>
-        </div>
-      </div>
-    );
+
+  useEffect(() => {
+    getTeammates(); // eslint-disable-next-line
+  }, []);
+
+  const getTeammates = async () => {
+    await Axios.get(URL)
+      .then(response => {
+        setTeammates(response.data);
+
+      });
+  };
+
+
+
+  const reversed = [];
+  const length = teammates.length;
+  for (let i = length - 1; i >= 0; i--) {
+    reversed.push(teammates[i]);
   }
-}
 
-export default Teammate;
+
+
+  return (
+    <div className="container">
+      {reversed.map((teammate, i) => (
+        <div key={i} className="card">
+          <img src={teammate.acf.image} alt="teammate" />
+          <div className="text">
+            <h4>{teammate.acf.name}</h4>
+            <p>{teammate.acf.position}</p>
+          </div>
+          <div className="social-links">
+            <a href={`mailto:${teammate.acf.email}`}>
+              <img
+                src="https://theshortcut.org/wp-content/uploads/2019/10/email-icon-01.png"
+                alt="mail-icon"
+              ></img>
+            </a>
+            <a href={teammate.acf.linkedin}>
+              {" "}
+              <img
+                src="https://theshortcut.org/wp-content/uploads/2019/10/linkedin-2.png"
+                alt="linkedin-icon"
+              />
+            </a>
+          </div>
+        </div>))}
+    </div>
+
+  );
+
+};
